Move composite pass to common place
