+++ mlir/CMakeLists.txt
@@ -56,6 +56,7 @@ set(SOURCES_LIST
     lib/Transforms/CastLowering.cpp
     lib/Transforms/CastUtils.cpp
     lib/Transforms/CommonOpts.cpp
+    lib/Transforms/CompositePass.cpp
     lib/Transforms/ConstUtils.cpp
     lib/Transforms/ExpandTuple.cpp
     lib/Transforms/FuncUtils.cpp
@@ -103,6 +104,7 @@ set(HEADERS_LIST
     include/imex/Transforms/CastLowering.hpp
     include/imex/Transforms/CastUtils.hpp
     include/imex/Transforms/CommonOpts.hpp
+    include/imex/Transforms/CompositePass.hpp
     include/imex/Transforms/ConstUtils.hpp
     include/imex/Transforms/ExpandTuple.hpp
     include/imex/Transforms/FuncUtils.hpp
